requirements.txt
apache-airflow
apache-airflow-providers-postgres
pandas
requests
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
nltk
textstat
feedparser
praw
yake
beautifulsoup4
newspaper3k
vaderSentiment
transformers
torch
unidecode
waybackpy
langchain
html2text
openai
lxml
lxml_html_clean
langchain_community


